requirements.txt
streamlit
deep-translator
python-pptx
pdf2docx
python-docx
opencv-python-headless
openpyxl
